[CHORE] #127 id값 타입 변경 Int -> Long

+++ app/src/main/java/com/starters/hsge/data/repository/ChatListRepositoryImpl.kt
@@ -9,7 +9,7 @@ import javax.inject.Inject
 class ChatListRepositoryImpl @Inject constructor(
     @RetrofitHSGE private val api: ChatRecordApi
 ) : ChatListRepository {
-    override suspend fun getChatList(roomId: Int): MessageInfo {
+    override suspend fun getChatList(roomId: Long): MessageInfo {
         return api.getChatRecord(roomId)
     }
 }
